Updated test and removed health tests
