+++ lumigator/python/mzai/backend/backend/tests/unit/api/routes/test_jobs.py
@@ -21,9 +21,9 @@ def test_get_job_status(
     json_ray_version,
     json_data_health_job_metadata_ray,
 ):
-    created_job = job_repository.create(name="test", description="")
+    created_job = job_repository.create(name="test", description="test desc")
 
-    # The Ray client will call the Ray API to get the version before gettting the job status
+    # The Ray client will call the Ray API to get the version before getting the job status
     # Mock the Ray version API
     request_mock.get(
         url=settings.RAY_VERSION_URL,
@@ -44,6 +44,11 @@ def test_get_job_status(
 
     data = response.json()
     assert data["status"].lower() == "running"
+    assert data["name"] == "test"
+    assert data["description"] == "test desc"
+    assert data["type"] == "SUBMISSION"
+    assert data["id"] == str(created_job.id)
+    assert data["driver_agent_http_address"] == "http://172.18.0.3:52365"
 
 
 def test_get_job_results(
@@ -54,24 +59,37 @@ def test_get_job_results(
     json_data_health_job_metadata_ray,
 ):
     created_job = job_repository.create(name="test", description="")
-
     expected_url_path = f"lumigator-storage/jobs/results/test/{created_job.id}/results.json"
-
     # The Ray client will call the Ray API to get the version during initialization
     request_mock.get(
         url=settings.RAY_VERSION_URL,
         status_code=status.HTTP_200_OK,
         text=json.dumps(load_json(json_ray_version)),
     )
-
     response = app_client.get(f"/jobs/{created_job.id}/result/download")
-
     assert response is not None
     assert response.status_code == status.HTTP_200_OK
     response_json = response.json()
     download_url = response_json["download_url"]
-
     match = re.search(r"lumigator-storage/.+?\.json", download_url)
     extracted_url_path = match.group(0) if match else None
-
     assert extracted_url_path == expected_url_path
+
+
+def test_job_logs(
+    app_client: TestClient,
+    request_mock,
+):
+    job_id = "d34dd34d-d34d-d34d-d34d-d34dd34dd34d"
+    log = "2024-11-13 02:00:08,889\\tINFO job_manager.py:530 -- Runtime env is setting up.\\n"
+    logs_content = f'{{"logs": "{log}"}}'
+
+    request_mock.get(
+        url=urllib.parse.urljoin(f"{settings.RAY_JOBS_URL}", f"{job_id}/logs"),
+        status_code=status.HTTP_200_OK,
+        text=logs_content,
+    )
+    response = app_client.get(f"/jobs/{job_id}/logs")
+    assert response is not None
+    assert response.status_code == status.HTTP_200_OK
+    assert json.loads(logs_content)["logs"] == json.loads(f'"{log}"')
